+++ lib/vm/src/instance/mod.rs
@@ -49,11 +49,6 @@ use wasmer_types::{
 pub type ImportInitializerFuncPtr =
     fn(*mut std::ffi::c_void, *const std::ffi::c_void) -> Result<(), *mut std::ffi::c_void>;
 
-/// This type holds thunks (delayed computations) for initializing the imported
-/// function's environments with the [`Instance`].
-pub(crate) type ImportInitializerThunks =
-    Vec<(Option<ImportInitializerFuncPtr>, *mut std::ffi::c_void)>;
-
 /// A WebAssembly instance.
 ///
 /// The type is dynamically-sized. Indeed, the `vmctx` field can
@@ -98,19 +93,99 @@ pub(crate) struct Instance {
     /// Handler run when `SIGBUS`, `SIGFPE`, `SIGILL`, or `SIGSEGV` are caught by the instance thread.
     pub(crate) signal_handler: Cell<Option<Box<SignalHandler>>>,
 
-    /// Functions to initialize the host environments in the imports
-    /// and pointers to the environments. These function pointers all
-    /// come from `WasmerEnv::init_with_instance`.
+    /// Functions to operate on host environments in the imports
+    /// and pointers to the environments.
     ///
     /// TODO: Be sure to test with serialize/deserialize and imported
     /// functions from other Wasm modules.
-    import_initializers: ImportInitializerThunks,
+    imported_function_envs: BoxedSlice<FunctionIndex, ImportFunctionEnv>,
 
     /// Additional context used by compiled WebAssembly code. This
     /// field is last, and represents a dynamically-sized array that
     /// extends beyond the nominal end of the struct (similar to a
     /// flexible array member).
     vmctx: VMContext,
+}
+
+/// A collection of data about host envs used by imported functions.
+#[derive(Debug)]
+pub enum ImportFunctionEnv {
+    /// The `vmctx` pointer does not refer to a host env, there is no
+    /// metadata about it.
+    NoEnv,
+    /// We're dealing with a user-defined host env.
+    ///
+    /// This host env may be either unwrapped (the user-supplied host env
+    /// directly) or wrapped. i.e. in the case of Dynamic functions, we
+    /// store our own extra data along with the user supplied env,
+    /// thus the `env` pointer here points to the outermost type.
+    Env {
+        /// The function environment. This is not always the user-supplied
+        /// env.
+        env: *mut std::ffi::c_void,
+        /// A clone function for duplicating the env.
+        clone: fn(*mut std::ffi::c_void) -> *mut std::ffi::c_void,
+        /// This field is not always present. When it is present, it
+        /// should be set to `None` after use to prevent double
+        /// initialization.
+        initializer: Option<ImportInitializerFuncPtr>,
+        /// The destructor to clean up the type in `env`.
+        ///
+        /// # Safety
+        /// - This function must be called ina synchronized way. For
+        ///   example, in the `Drop` implementation of this type.
+        destructor: unsafe fn(*mut std::ffi::c_void),
+    },
+}
+
+impl ImportFunctionEnv {
+    /// Get the `initializer` function pointer if it exists.
+    fn initializer(&self) -> Option<ImportInitializerFuncPtr> {
+        match self {
+            Self::Env { initializer, .. } => *initializer,
+            _ => None,
+        }
+    }
+}
+
+impl Clone for ImportFunctionEnv {
+    fn clone(&self) -> Self {
+        match &self {
+            Self::NoEnv => Self::NoEnv,
+            Self::Env {
+                env,
+                clone,
+                destructor,
+                initializer,
+            } => {
+                let new_env = (*clone)(*env);
+                Self::Env {
+                    env: new_env,
+                    clone: *clone,
+                    destructor: *destructor,
+                    initializer: *initializer,
+                }
+            }
+        }
+    }
+}
+
+impl Drop for ImportFunctionEnv {
+    fn drop(&mut self) {
+        match self {
+            ImportFunctionEnv::Env {
+                env, destructor, ..
+            } => {
+                // # Safety
+                // - This is correct because we know no other references
+                //   to this data can exist if we're dropping it.
+                unsafe {
+                    (destructor)(*env);
+                }
+            }
+            ImportFunctionEnv::NoEnv => (),
+        }
+    }
 }
 
 impl fmt::Debug for Instance {
@@ -159,7 +234,7 @@ impl Instance {
         &self,
         index: FunctionIndex,
     ) -> Option<ImportInitializerFuncPtr> {
-        self.import_initializers[index.as_u32() as usize].0
+        self.imported_function_envs[index].initializer()
     }
 
     /// Return a pointer to the `VMFunctionImport`s.
@@ -692,27 +767,18 @@ impl Instance {
     }
 }
 
-/// TODO: update these docs
-/// An `InstanceRef` is responsible to allocate, to deallocate,
+/// An `InstanceRef` is responsible to properly deallocate,
 /// and to give access to an `Instance`, in such a way that `Instance`
 /// is unique, can be shared, safely, across threads, without
 /// duplicating the pointer in multiple locations. `InstanceRef`
 /// must be the only “owner” of an `Instance`.
 ///
 /// Consequently, one must not share `Instance` but
-/// `InstanceRef`. It acts like an Atomically Reference Counted
+/// `InstanceRef`. It acts like an Atomically Reference Counter
 /// to `Instance`. In short, `InstanceRef` is roughly a
 /// simplified version of `std::sync::Arc`.
 ///
-/// It is important to remind that `Instance` is dynamically-sized
-/// based on `VMOffsets`: The `Instance.vmctx` field represents a
-/// dynamically-sized array that extends beyond the nominal end of the
-/// type. So in order to create an instance of it, we must:
-///
-/// 1. Define the correct layout for `Instance` (size and alignment),
-/// 2. Allocate it properly.
-///
-/// This allocation must be freed with [`InstanceRef::deallocate_instance`]
+/// This `InstanceRef` must be freed with [`InstanceRef::deallocate_instance`]
 /// if and only if it has been set correctly. The `Drop` implementation of
 /// [`InstanceRef`] calls its `deallocate_instance` method without
 /// checking if this  property holds, only when `Self.strong` is equal to 1.
@@ -752,20 +818,17 @@ pub struct InstanceRef {
 }
 
 impl InstanceRef {
-    /// Create a new `InstanceRef`. It allocates nothing. It
-    /// fills nothing. The `Instance` must be already valid and
-    /// filled. `self_ptr` and `self_layout` must be the pointer and
-    /// the layout returned by `Self::allocate_self` used to build
-    /// `Self`.
+    /// Create a new `InstanceRef`. It allocates nothing. It fills
+    /// nothing. The `Instance` must be already valid and
+    /// filled.
     ///
     /// # Safety
     ///
     /// `instance` must a non-null, non-dangling, properly aligned,
     /// and correctly initialized pointer to `Instance`. See
-    /// [`InstanceAllocator::new`] for an example of how to correctly use
+    /// [`InstanceAllocator`] for an example of how to correctly use
     /// this API.
-    /// TODO: update docs
-    pub(crate) unsafe fn new(instance: NonNull<Instance>, instance_layout: Layout) -> Self {
+    pub(self) unsafe fn new(instance: NonNull<Instance>, instance_layout: Layout) -> Self {
         Self {
             strong: Arc::new(atomic::AtomicUsize::new(1)),
             instance_layout,
@@ -917,12 +980,12 @@ impl Drop for InstanceRef {
 /// providing useful higher-level API.
 #[derive(Debug, PartialEq)]
 pub struct InstanceHandle {
-    /// The `InstanceRef`. See its documentation to learn more.
+    /// The [`InstanceRef`]. See its documentation to learn more.
     instance: InstanceRef,
 }
 
 impl InstanceHandle {
-    /// Create a new `InstanceHandle` pointing at a new `InstanceRef`.
+    /// Create a new `InstanceHandle` pointing at a new [`InstanceRef`].
     ///
     /// # Safety
     ///
@@ -955,7 +1018,7 @@ impl InstanceHandle {
         imports: Imports,
         vmshared_signatures: BoxedSlice<SignatureIndex, VMSharedSignatureIndex>,
         host_state: Box<dyn Any>,
-        import_initializers: ImportInitializerThunks,
+        imported_function_envs: BoxedSlice<FunctionIndex, ImportFunctionEnv>,
     ) -> Result<Self, Trap> {
         let vmctx_globals = finished_globals
             .values()
@@ -979,14 +1042,14 @@ impl InstanceHandle {
                 passive_data,
                 host_state,
                 signal_handler: Cell::new(None),
-                import_initializers,
+                imported_function_envs,
                 vmctx: VMContext {},
             };
 
-            let instance_allocator = allocator.write_instance(instance);
+            let instance_ref = allocator.write_instance(instance);
 
             Self {
-                instance: instance_allocator,
+                instance: instance_ref,
             }
         };
         let instance = handle.instance().as_ref();
@@ -1131,7 +1194,7 @@ impl InstanceHandle {
                     signature,
                     vmctx,
                     call_trampoline,
-                    instance_allocator: Some(instance),
+                    instance_ref: Some(instance),
                 }
                 .into()
             }
@@ -1144,7 +1207,7 @@ impl InstanceHandle {
                 };
                 VMExportTable {
                     from,
-                    instance_allocator: Some(instance),
+                    instance_ref: Some(instance),
                 }
                 .into()
             }
@@ -1157,7 +1220,7 @@ impl InstanceHandle {
                 };
                 VMExportMemory {
                     from,
-                    instance_allocator: Some(instance),
+                    instance_ref: Some(instance),
                 }
                 .into()
             }
@@ -1172,7 +1235,7 @@ impl InstanceHandle {
                 };
                 VMExportGlobal {
                     from,
-                    instance_allocator: Some(instance),
+                    instance_ref: Some(instance),
                 }
                 .into()
             }
@@ -1267,19 +1330,26 @@ impl InstanceHandle {
     ) -> Result<(), Err> {
         let instance_ref = self.instance.as_mut();
 
-        for (func, env) in instance_ref.import_initializers.drain(..) {
-            if let Some(ref f) = func {
-                // transmute our function pointer into one with the correct error type
-                let f = mem::transmute::<
-                    &ImportInitializerFuncPtr,
-                    &fn(*mut ffi::c_void, *const ffi::c_void) -> Result<(), Err>,
-                >(f);
-                f(env, instance_ptr)?;
+        for import_function_env in instance_ref.imported_function_envs.values_mut() {
+            match import_function_env {
+                ImportFunctionEnv::Env {
+                    env,
+                    ref mut initializer,
+                    ..
+                } => {
+                    if let Some(f) = initializer {
+                        // transmute our function pointer into one with the correct error type
+                        let f = mem::transmute::<
+                            &ImportInitializerFuncPtr,
+                            &fn(*mut ffi::c_void, *const ffi::c_void) -> Result<(), Err>,
+                        >(f);
+                        f(*env, instance_ptr)?;
+                    }
+                    *initializer = None;
+                }
+                ImportFunctionEnv::NoEnv => (),
             }
         }
-        // free memory now that it's empty.
-        instance_ref.import_initializers.shrink_to_fit();
-
         Ok(())
     }
 }